# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Nyanbin API
  version: '1'
  summary: Zero-knowledge encrypted note sharing.
  description: |
    Nyanbin is a zero-knowledge pastebin. All encryption and decryption happens
    on the client; the server only ever stores an **opaque envelope** and can
    never read note contents or passwords. Delete tokens are transient bearer
    capabilities: the server generates or receives them when required but
    never logs or stores them; only their SHA-256 digests are persisted.

    ## Creating a note: reserve → encrypt → commit

    1. `POST /api/notes/reserve` with the desired lifecycle. The server mints a
       note ID, a `deleteToken` capability, and the authoritative lifecycle
       (`expiresAt` in milliseconds since the Unix epoch, optional `maxReads`).
    2. The client encrypts the note **locally** into a protocol-v1 envelope and
       binds the returned lifecycle into it.
    3. `PUT /api/notes/{id}` commits the envelope together with the **same**
       lifecycle and the SHA-256 hex digest of the delete token.

    A reservation lives for the instance-configured reservation TTL (10–600
    seconds; 120 seconds by default). If it is not committed in time it
    silently expires and the ID is never published.

    ## Idempotency and retries

    Reserve is **not idempotent**: a lost response is harmless (the unused
    reservation simply expires) and the correct retry is a fresh reserve call.
...

    ## The envelope is opaque

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security: []
...
paths:
...
  /api/notes/{id}:
    parameters:
      - $ref: '#/components/parameters/NoteId'
    put:
      operationId: commitNote
      tags: [notes]
      summary: Commit an encrypted envelope to a reservation
      description: |
        Publishes the note by attaching the locally encrypted envelope to a
        prior reservation. The lifecycle **must equal** the lifecycle returned
        by reserve, and `deleteTokenHash` must be the SHA-256 hex digest of
        the reserved delete token. Commit transmits only that digest, not the
        raw token.

        The envelope is opaque canonical base64url produced by the published
        protocol implementation (npm package `nyanbin`, export
        `./shared/protocol`). Never send plaintext.

        Commit is safe against duplicate delivery: a second commit to the same
        ID yields `409 reservation_mismatch`, so at most one envelope is ever
        published per reservation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitRequest'
      responses:
        '201':
          description: Note published.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResponse'
        '400':
          description: >
            Malformed request (`invalid_request`, `invalid_json`), malformed
            ID (`invalid_id`), lifecycle mismatch or out of range
            (`invalid_lifecycle`), or non-canonical envelope
            (`invalid_envelope`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Reservation expired or unknown (`reservation_not_found`). Reserve again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: >
            Commit does not match the reservation, or an envelope was already
            committed for this ID (`reservation_mismatch`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Envelope exceeds the instance size limit (`payload_too_large`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          description: >
            Writes temporarily paused by the operator (`writes_disabled`) or
            backing store unreachable (`storage_unavailable`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '507':
          description: >
            Storage pressure brownout (`storage_pressure`): near the storage
            budget only small commits are accepted; at the budget all commits
            are rejected. Temporary; retry later or with a smaller note.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      operationId: getNoteInfo
      tags: [notes]
      summary: Note metadata (non-consuming)
      description: |
        Returns lifecycle metadata and whether the note is password protected
        **without** consuming a read. Use this to inspect `remainingReads`
        before revealing.
      responses:
        '200':
          description: Note metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoResponse'
        '400':
          description: Malformed note ID (`invalid_id`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Note unknown, expired, or fully consumed (`note_not_found`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          description: Backing store unreachable (`storage_unavailable`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note with its delete token
      description: |
        Deletes the note when the presented `deleteToken` matches the SHA-256
        digest stored at commit time. The token is a bearer capability: the
        server never stores it in the clear.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteRequest'
      responses:
        '204':
          description: Note deleted.
        '400':
          description: Malformed note ID (`invalid_id`) or body (`invalid_json`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Delete token is malformed or does not match (`invalid_delete_token`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Note unknown, expired, or fully consumed (`note_not_found`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large (`payload_too_large`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          description: Backing store unreachable (`storage_unavailable`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    NoteId:
      name: id
      in: path
      required: true
      description: Note ID (32 base62 characters).
      schema:
        type: string
        pattern: '^[0-9A-Za-z]{32}$'
        minLength: 32
        maxLength: 32
  headers:
    RetryAfter:
      description: Seconds to wait before retrying (integer, at least 1).
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
  responses:
    RateLimited:
      description: >
        Too many requests (`rate_limited`). Also returned when the per-client
        hourly upload byte quota is exhausted; then `Retry-After` points at
        the next hour boundary.
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Lifecycle:
      type: object
      description: >
        Authoritative note lifecycle. `maxReads` is a positive count or
        absent for unlimited reads — never 0.
      additionalProperties: false
      required: [expiresAt]
      properties:
        expiresAt:
          type: integer
          format: int64
          minimum: 0
          description: Expiry as milliseconds since the Unix epoch (uint64).
        maxReads:
          type: integer
          format: uint32
          minimum: 1
          maximum: 4294967295
          description: Maximum read count; absent means unlimited. Never 0.
    InfoLifecycle:
      type: object
      description: Lifecycle as reported by the info endpoint.
      additionalProperties: false
      required: [expiresAt]
      properties:
        expiresAt:
          type: integer
          format: int64
          minimum: 0
          description: Expiry as milliseconds since the Unix epoch (uint64).
        maxReads:
          type: integer
          format: uint32
          minimum: 1
          maximum: 4294967295
          description: Maximum read count; absent means unlimited.
        remainingReads:
          type: integer
          format: uint32
          minimum: 0
          maximum: 4294967295
          description: Reads left before the note is consumed; absent when unlimited.
...
    CommitRequest:
      type: object
      description: Envelope commit for a prior reservation.
      additionalProperties: false
      required: [protocol, envelope, lifecycle, deleteTokenHash]
      properties:
        protocol:
          type: integer
          const: 1
          description: Wire protocol version. Must be 1.
        envelope:
          type: string
          pattern: '^[A-Za-z0-9_-]+$'
          description: >
            Opaque canonical base64url envelope produced by the published
            protocol implementation (npm package `nyanbin`, export
            `./shared/protocol`). Never plaintext.
        lifecycle:
          $ref: '#/components/schemas/Lifecycle'
        deleteTokenHash:
          type: string
          pattern: '^[0-9a-f]{64}$'
          minLength: 64
          maxLength: 64
          description: SHA-256 of the reserved delete token, lowercase hex.
        passwordProtected:
          type: boolean
          default: false
          description: Whether the envelope additionally requires a password to decrypt.
    CreateResponse:
      type: object
      description: Confirmation that the note was published.
      additionalProperties: false
      required: [id]
      properties:
        id:
          type: string
          pattern: '^[0-9A-Za-z]{32}$'
          minLength: 32
          maxLength: 32
          description: Note ID (32 base62 characters).
    InfoResponse:
      type: object
      description: Non-consuming note metadata.
      additionalProperties: false
      required: [protocol, lifecycle, passwordProtected]
      properties:
        protocol:
          type: integer
          const: 1
          description: Wire protocol version.
        lifecycle:
          $ref: '#/components/schemas/InfoLifecycle'
        passwordProtected:
          type: boolean
          description: Whether decryption additionally requires a password.
...
    DeleteRequest:
      type: object
      description: Delete capability presentation.
      additionalProperties: false
      required: [deleteToken]
      properties:
        deleteToken:
          type: string
          pattern: '^[A-Za-z0-9_-]{43}$'
          minLength: 43
          maxLength: 43
          description: The delete token returned by reserve.
...
    Error:
      type: object
      description: >
        Uniform error body. `code` is machine-readable and append-only across
        the v1 surface; `message` is human-readable and may change.
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - invalid_json
            - payload_too_large
            - invalid_id
            - invalid_lifecycle
            - invalid_envelope
            - invalid_delete_token
            - invalid_short_code
            - reservation_not_found
            - reservation_mismatch
            - note_not_found
            - route_not_found
            - method_not_allowed
            - rate_limited
            - short_link_requires_password
            - short_code_space_exhausted
            - id_space_exhausted
            - storage_unavailable
            - storage_pressure
            - writes_disabled
            - short_disabled
            - internal_error
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable explanation. Not part of the stable surface.